+++ src/main/java/com/iemr/hwc/repo/masterrepo/anc/PersonalHabitTypeRepo.java
@@ -34,7 +34,7 @@ import com.iemr.hwc.data.masterdata.anc.PersonalHabitType;
 
 @Repository
 @RestResource(exported = false)
-public interface PersonalHabitTypeRepo extends CrudRepository<PersonalHabitType, Long>{
+public interface PersonalHabitTypeRepo extends CrudRepository<PersonalHabitType, Short>{
 	
 	@Query("select personalHabitTypeID, habitType, habitValue from PersonalHabitType where deleted = false and habitType=:habitType "
 			+ "order by habitValue ")
