+++ src/main/java/com/iemr/hwc/repo/masterrepo/nurse/VisitReasonMasterRepo.java
@@ -31,7 +31,7 @@ import org.springframework.stereotype.Repository;
 import com.iemr.hwc.data.masterdata.nurse.VisitReason;
 @Repository
 @RestResource(exported = false)
-public interface VisitReasonMasterRepo extends CrudRepository<VisitReason, Long>{
+public interface VisitReasonMasterRepo extends CrudRepository<VisitReason, Short>{
 
 	@Query("select visitReasonID, visitReason from VisitReason where deleted = false order by visitReason ")
 	public ArrayList<Object[]> getVisitReasonMaster();
